GoogleAuth, HashPasswords, ConfirmationCode, CRON
The logic of working with authorization via Google has been brought to the frontend.
Adjusted the length of the password field in the Pydantic and database models in order to increase security.
The logic of sending the confirmation code has been revised.
The function of deleting outdated information from the database has been improved.